fixing healthcheck which was breaking due to jsonify error
The jsonify call was erroring with the message "'Request' object has no attribute 'is_xhr'" - potentially due to recent deprecation.

This causes the / page to return a 500 error, which was causing the ALB in the tutorial to fail. have replaced with output that will not fail the check, and produces the same message.